feat(core): handle documentation create event
