+++ packages/core/src/scd-xml/index.ts
@@ -1,1 +1,2 @@
 export * from './foundation'
+export * from './utils.scd-xml'
